### Fixed
- Error handling versions for metadata documents (issues #55)
### Changed
- Building service all tests now enabled by default (use -Dprofile=minimal to execute only mandatory tests)

+++ src/main/java/edu/kit/datamanager/metastore2/util/MetadataRecordUtil.java
@@ -153,6 +153,7 @@ public class MetadataRecordUtil {
     // Create additional metadata record for faster access
     DataRecord dataRecord = new DataRecord();
     dataRecord.setMetadataId(createResource.getId());
+    dataRecord.setVersion(record.getRecordVersion());
     dataRecord.setSchemaId(record.getSchema().getIdentifier());
     dataRecord.setMetadataDocumentUri(contentInformation.getContentUri());
     dataRecord.setDocumentHash(contentInformation.getHash());
@@ -205,6 +206,7 @@ public class MetadataRecordUtil {
       dataResource = DataResourceUtils.copyDataResource(dataResource);
     }
 
+    boolean noChanges = true;
     if (document != null) {
       record = migrateToMetadataRecord(applicationProperties, dataResource, false);
       validateMetadataDocument(applicationProperties, record, document);
@@ -212,7 +214,6 @@ public class MetadataRecordUtil {
       ContentInformation info;
       info = getContentInformationOfResource(applicationProperties, dataResource);
       // Check for changes...
-      boolean noChanges = true;
       try {
         byte[] currentFileContent;
         File file = new File(URI.create(info.getContentUri()));
@@ -238,11 +239,13 @@ public class MetadataRecordUtil {
         if (version != null) {
           dataResource.setVersion(Long.toString(Long.parseLong(version) + 1l));
         }
-        ContentInformation addFile = ContentDataUtils.addFile(applicationProperties, dataResource, document, info.getRelativePath(), null, true, supplier);
-        Optional<DataRecord> dataRecord = dataRecordDao.findByMetadataId(dataResource.getId());
-        if (dataRecord.isPresent()) {
-          dataRecordDao.delete(dataRecord.get());
-        }
+        ContentDataUtils.addFile(applicationProperties, dataResource, document, info.getRelativePath(), null, true, supplier);
+      }
+    }
+    if (noChanges) {
+      Optional<DataRecord> dataRecord = dataRecordDao.findTopByMetadataIdOrderByVersionDesc(dataResource.getId());
+      if (dataRecord.isPresent()) {
+        dataRecordDao.delete(dataRecord.get());
       }
     }
     dataResource = DataResourceUtils.updateResource(applicationProperties, resourceId, dataResource, eTag, supplier);
@@ -255,9 +258,10 @@ public class MetadataRecordUtil {
           String eTag,
           Function<String, String> supplier) {
     DataResourceUtils.deleteResource(applicationProperties, id, eTag, supplier);
-    Optional<DataRecord> dataRecord = dataRecordDao.findByMetadataId(id);
-    if (dataRecord.isPresent()) {
+    Optional<DataRecord> dataRecord = dataRecordDao.findTopByMetadataIdOrderByVersionDesc(id);
+    while (dataRecord.isPresent()) {
       dataRecordDao.delete(dataRecord.get());
+      dataRecord = dataRecordDao.findTopByMetadataIdOrderByVersionDesc(id);
     }
   }
 
@@ -399,20 +403,18 @@ public class MetadataRecordUtil {
       }
       DataRecord dataRecord = null;
       long nano2 = System.nanoTime() / 1000000;
-      Optional<DataRecord> dataRecordResult = dataRecordDao.findByMetadataId(dataResource.getId());
+      Optional<DataRecord> dataRecordResult = dataRecordDao.findByMetadataIdAndVersion(dataResource.getId(), recordVersion);
       long nano3 = System.nanoTime() / 1000000;
       long nano4 = nano3;
       boolean isAvailable = false;
       boolean saveDataRecord = false;
       if (dataRecordResult.isPresent()) {
         LOG.trace("Get document URI from DataRecord.");
-        if (dataRecordResult.get().getVersion() == recordVersion) {
         dataRecord = dataRecordResult.get();
         nano4 = System.nanoTime() / 1000000;
         metadataRecord.setMetadataDocumentUri(dataRecord.getMetadataDocumentUri());
         metadataRecord.setDocumentHash(dataRecord.getDocumentHash());
-        isAvailable = true;        
-        }
+        isAvailable = true;
       } else {
         saveDataRecord = true;
       }
@@ -803,6 +805,7 @@ public class MetadataRecordUtil {
       LOG.trace("Transform to data record!");
       dataRecord = new DataRecord();
       dataRecord.setMetadataId(result.getId());
+      dataRecord.setVersion(result.getRecordVersion());
       dataRecord.setSchemaId(result.getSchema().getIdentifier());
       dataRecord.setDocumentHash(result.getDocumentHash());
       dataRecord.setMetadataDocumentUri(result.getMetadataDocumentUri());
